feat: integrate calcore for Delta E and white-balance calibration
- Implemented Issue #6 using calcore for active Delta E path.
- Updated calibrator/session.py and calibrator/quality.py with new 1.5/3.0 thresholds.
- Refactored calibrator module: deleted colour.py and split helpers into utils.py, models.py, and runtime.py.
- Expanded calcore/spaces.py to cover BT.709, P3 D65, and BT.2020.
- Added migration tests and updated legacy test suite.
- Verified: 432 tests passed.
