+++ calcore/spaces.py
@@ -1,6 +1,24 @@
 from __future__ import annotations
 
-from typing import Sequence, Tuple
+from typing import Dict, Sequence, Tuple
+
+BT709_PRIMARIES = {
+    "red": (0.640, 0.330),
+    "green": (0.300, 0.600),
+    "blue": (0.150, 0.060),
+}
+
+P3D65_PRIMARIES = {
+    "red": (0.680, 0.320),
+    "green": (0.265, 0.690),
+    "blue": (0.150, 0.060),
+}
+
+BT2020_PRIMARIES = {
+    "red": (0.708, 0.292),
+    "green": (0.170, 0.797),
+    "blue": (0.131, 0.046),
+}
 
 BT709_RGB_TO_XYZ = (
     (0.4123908, 0.35758434, 0.18048079),
@@ -34,10 +52,21 @@ def rgb_to_xyz(
 
 def detect_matrix(name: str):
     key = name.lower()
-    if key in ("bt709", "709", "rec709"):
+    if key in ("bt709", "709", "rec709", "rec.709", "rec 709"):
         return BT709_RGB_TO_XYZ
-    if key in ("p3d65", "displayp3", "p3"):
+    if key in ("p3d65", "displayp3", "p3", "p3 d65"):
         return P3D65_RGB_TO_XYZ
-    if key in ("bt2020", "2020", "rec2020"):
+    if key in ("bt2020", "2020", "rec2020", "rec.2020", "rec 2020"):
         return BT2020_RGB_TO_XYZ
+    raise ValueError(f"Unknown target space: {name}")
+
+
+def detect_primaries(name: str) -> Dict[str, Tuple[float, float]]:
+    key = name.lower()
+    if key in ("bt709", "709", "rec709", "rec.709", "rec 709"):
+        return dict(BT709_PRIMARIES)
+    if key in ("p3d65", "displayp3", "p3", "p3 d65"):
+        return dict(P3D65_PRIMARIES)
+    if key in ("bt2020", "2020", "rec2020", "rec.2020", "rec 2020"):
+        return dict(BT2020_PRIMARIES)
     raise ValueError(f"Unknown target space: {name}")
