+++ calibrator/__init__.py
@@ -1,23 +1,31 @@
 """
 ZRO Calibration Helper package.
 
-All public symbols are re-exported here.
-Import from the submodules directly for cleaner dependency graphs.
+Most colour-science primitives now come from `calcore`; this package re-exports
+the stable pieces the FastAPI backend and tests use.
 """
 
-from .colour import (
-    REQUIRED_PACKAGES, ensure_packages, console,
-    D65_XY, D65_XYZ,
-    REC709_PRIMARIES, P3_PRIMARIES, REC2020_PRIMARIES,
-    GRAYSCALE_LEVELS, SATURATION_LEVELS, COLOUR_TARGETS,
-    GAMMA_SDR, GAMMA_HDR_PQ,
-    CalMode, Measurement, CalibrationTarget, CalibrationReport,
-    SDR_TARGET, HDR10_TARGET, DV_TARGET,
-    delta_e_cie76, xyY_to_XYZ, XYZ_to_lab, xyY_to_lab,
-    delta_xy, gamma_from_luminance, rating_emoji, direction_hint,
-    stimulus_pct_from_code_value,
+from calcore import (
+    BT2020_PRIMARIES as REC2020_PRIMARIES,
+    BT709_PRIMARIES as REC709_PRIMARIES,
+    CalMode,
+    CalibrationTarget,
+    D65_XY,
+    D65_XYZ,
+    DV_TARGET,
+    HDR10_TARGET,
+    Measurement,
+    P3D65_PRIMARIES as P3_PRIMARIES,
+    SDR_TARGET,
+    XYZ_to_lab,
+    ciede2000,
+    delta_e_cie76,
+    xyY_to_XYZ,
+    xyY_to_lab,
 )
+from .models import CalibrationReport
 from .profiles import TVProfile, TV_PROFILES, DEFAULT_TV_PROFILE
+from .runtime import REQUIRED_PACKAGES, console, ensure_packages
 from .guidance import (
     cms_control_plan,
     cms_hints,
@@ -34,7 +42,21 @@ from .guidance import (
 from .quality import step_quality
 from .reports import report_payload, render_report_html
 from .session import SessionStore
+from .utils import (
+    delta_e_ciede2000_xyY,
+    delta_xy,
+    direction_hint,
+    gamma_from_luminance,
+    rating_emoji,
+    stimulus_pct_from_code_value,
+)
 from .zro_import import parse_zro_csv, merge_into_session, ZROImportResult
+
+GRAYSCALE_LEVELS = list(range(0, 101, 5))
+SATURATION_LEVELS = [25, 50, 75, 100]
+COLOUR_TARGETS = ["red", "green", "blue", "cyan", "magenta", "yellow"]
+GAMMA_SDR = 2.2
+GAMMA_HDR_PQ = "PQ"
 
 __all__ = [
     "REQUIRED_PACKAGES", "ensure_packages", "console",
@@ -44,7 +66,7 @@ __all__ = [
     "GAMMA_SDR", "GAMMA_HDR_PQ",
     "CalMode", "Measurement", "CalibrationTarget", "CalibrationReport",
     "SDR_TARGET", "HDR10_TARGET", "DV_TARGET",
-    "delta_e_cie76", "xyY_to_XYZ", "XYZ_to_lab", "xyY_to_lab",
+    "ciede2000", "delta_e_cie76", "delta_e_ciede2000_xyY", "xyY_to_XYZ", "XYZ_to_lab", "xyY_to_lab",
     "delta_xy", "gamma_from_luminance", "rating_emoji", "direction_hint",
     "stimulus_pct_from_code_value",
     "TVProfile", "TV_PROFILES", "DEFAULT_TV_PROFILE",
